fix : 퀴즈 api 수정 및 오늘의 퀴즈 api 추가

+++ quiz-service/src/main/java/com/exit/quiz/domain/repository/QuizAttemptsRepository.java
@@ -19,7 +19,7 @@ public interface QuizAttemptsRepository extends JpaRepository<QuizAttempts, Long
                     from QuizAttempts qa join qa.quiz q join q.quizCategory qc
                     where qc.id in(:categoryIdList) and qa.userId = :userId
             """)
-    Slice<GetSolvedQuiz> findByQuizCategoryIdIn(@Param("categoryIdList") List<Long> categoryIdList,
+    Slice<GetSolvedQuiz> findByQuizCategoryIdIn(@Param("categoryIdList") List<Short> categoryIdList,
                                                 @Param("userId") Long userId,
                                                 Pageable pageable
     );
